% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{browse_website_rnaseq_search}
\alias{browse_website_rnaseq_search}
\title{Browse GEO search website for RNA-seq datasets}
\usage{
browse_website_rnaseq_search()
}
\description{
This function opens a browser window to the NCBI GEO website
with a search for RNA-seq datasets. It is included as a convenience
function to remind users of how to search for RNA-seq datasets using
the NCBI GEO website and an "rnaseq counts" filter.
}
\examples{
\dontrun{
browse_website_rnaseq_search()
}
}
