% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browseGEOAccession.R
\name{browseGEOAccession}
\alias{browseGEOAccession}
\title{Open the GEO page for a given accession}
\usage{
browseGEOAccession(geo)
}
\arguments{
\item{geo}{A GEO accession number}
}
\description{
Sometimes, you just need to see the GEO website page for a
GEO accession. This function opens the GEO page for a given
accession number in the default browser.
}
\examples{
\dontrun{
browseGEOAccession("GSE262484")
}

}
\seealso{
\code{\link{urlForAccession}}
}
