% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{clearGEOCache}
\alias{clearGEOCache}
\title{Clear the GEOquery download cache}
\usage{
clearGEOCache()
}
\value{
\code{NULL}, invisibly.
}
\description{
Remove all entries from the persistent download cache (see
\code{\link{geoCache}}).
}
\seealso{
\code{\link{geoCache}}
}
