\name{Generic functions}
\alias{dataTable}
\alias{Accession}
\alias{Columns}
\alias{GPLList}
\alias{GSMList}
\alias{Meta}
\alias{Table}
\title{Generic functions for GEOquery}
\description{
  The main documentation is in the Class documentation
}
\author{Sean Davis}
\seealso{\code{\link{GEOData-class}}}

\keyword{ IO }
