% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGEOfile.R
\name{downloadFile}
\alias{downloadFile}
\title{Wrapper for download.file() that removes files upon download failure}
\usage{
downloadFile(url, destfile, mode, quiet = TRUE)
}
\description{
Wrapper for download.file() that removes files upon download failure
}
