% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{extractFilenameFromDownloadURL}
\alias{extractFilenameFromDownloadURL}
\title{Extract filename from a GEO download URL}
\usage{
extractFilenameFromDownloadURL(url)
}
\arguments{
\item{url}{A GEO download URL}
}
\value{
A character vector with the filename
}
\description{
This function extracts the filename from a GEO download URL.
The filename is expected to be a query parameter called "file".
If the query parameter is not found, the function returns NULL.
}
\details{
The idea is to use this function to extract filenames that
contain important metadata from the GEO RNA-seq quantification.

In particular, the filename is expected to contain the genome build
and species information that we can attach to the SummarizedExperiment.
}
\keyword{internal}
