% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEOquery-package.R
\docType{class}
\name{GDS-class}
\alias{GDS-class}
\title{Class 'GDS'}
\description{
A class describing a GEO GDS entity
}
\section{Objects from the Class}{
 Objects of this class are returned by
\code{\link{getGEO}}; they are not normally constructed directly.
}

\seealso{
\code{\link{GEOData-class}}
}
\author{
Sean Davis
}
\keyword{classes}
