% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{geoCache}
\alias{geoCache}
\title{GEOquery download cache}
\usage{
geoCache()
}
\value{
A \code{BiocFileCache} object.
}
\description{
Return the \code{BiocFileCache} object that backs GEOquery's persistent
download cache (see \code{\link{clearGEOCache}}). The cache is used by the
download functions only when \code{options(GEOquery.cache = TRUE)} is set;
its location defaults to \code{tools::R_user_dir("GEOquery", "cache")} and can
be overridden with \code{options(GEOquery.cache.path = ...)}.
}
\seealso{
\code{\link{clearGEOCache}}
}
