% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEOquery-package.R
\docType{class}
\name{GEOData-class}
\alias{GEOData-class}
\alias{Accession,GEOData-method}
\alias{Columns,GEOData-method}
\alias{Meta,GEOData-method}
\alias{Table,GEOData-method}
\alias{dataTable,GEOData-method}
\alias{show,GEOData-method}
\title{Class 'GEOData'}
\description{
A virtual class for holding GEO samples, platforms, and datasets
}
\section{Objects from the Class}{
 Objects of this class are returned by
\code{\link{getGEO}}; they are not normally constructed directly.
}

\seealso{
\code{\link{GDS-class}}, \code{\link{GPL-class}},
\code{\link{GSM-class}}, \code{\link{GEODataTable-class}},
}
\author{
Sean Davis
}
\keyword{classes}
