% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{get_rna_quant_annotation_link}
\alias{get_rna_quant_annotation_link}
\title{Get the RNA-seq quantification annotation link}
\usage{
get_rna_quant_annotation_link(links)
}
\arguments{
\item{links}{A geoDownloadLinks object}
}
\value{
A character vector with the link to the annotation file
}
\description{
This function extracts the link to the RNA-seq quantification annotation
file from a geoDownloadLinks object.
}
\keyword{internal}
