% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGEOSuppFiles.R
\name{getDirListing}
\alias{getDirListing}
\title{get a directory listing from NCBI GEO}
\usage{
getDirListing(url)
}
\description{
This one makes some assumptions about the structure of the HTML response
returned.
}
