% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGEOSuppFiles.R
\name{getGEOSuppFileURL}
\alias{getGEOSuppFileURL}
\title{Get GEO supplemental file URL for a given GEO accession}
\usage{
getGEOSuppFileURL(GEO)
}
\arguments{
\item{GEO}{A GEO accession, e.g. a Series ("GSE..."), Sample ("GSM..."), or
Platform ("GPL...") id. The accession type determines which FTP \verb{suppl/}
directory URL is returned.}
}
\value{
A character(1) URL of the accession's supplementary-file directory on
the NCBI GEO FTP site.
}
\description{
Get GEO supplemental file URL for a given GEO accession
}
\examples{
# an example of a GEO supplemental file URL
# with a set of single-cell RNA-seq data
url = getGEOSuppFileURL("GSE161228")
url

\dontrun{
  browseURL(url)
}

}
