% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{getRNAQuantRawCountsURL}
\alias{getRNAQuantRawCountsURL}
\title{Get the link to the raw counts file from GEO}
\usage{
getRNAQuantRawCountsURL(links)
}
\arguments{
\item{links}{A geoDownloadLinks object}
}
\value{
A character vector with the link to the raw counts file
}
\description{
This function extracts the link to the raw counts file from a
geoDownloadLinks object.
}
\keyword{internal}
