% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEOquery-package.R
\docType{class}
\name{GPL-class}
\alias{GPL-class}
\alias{GPL}
\alias{GPL,GDS-method}
\title{Class 'GPL'}
\description{
Contains a full GEO Platform entity
}
\section{Objects from the Class}{
 Objects of this class are returned by
\code{\link{getGEO}}; they are not normally constructed directly.
}

\seealso{
\code{\link{GEOData-class}}
}
\author{
Sean Davis
}
\keyword{classes}
