% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEOquery-package.R
\docType{class}
\name{GSM-class}
\alias{GSM-class}
\title{Class 'GSM'}
\description{
A class containing a GEO Sample entity
}
\section{Objects from the Class}{
 Objects of this class are returned by
\code{\link{getGEO}}; they are not normally constructed directly.
}

\seealso{
\code{\link{GEOData-class}}
}
\author{
Sean Davis
}
\keyword{classes}
