% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{hasRNASeqQuantifications}
\alias{hasRNASeqQuantifications}
\title{Does a GEO accession have RNA-seq quantifications?}
\usage{
hasRNASeqQuantifications(accession)
}
\arguments{
\item{accession}{GEO accession number}
}
\value{
TRUE if the GEO accession number has RNA-seq quantifications
available, FALSE otherwise.
}
\description{
This function checks if a GEO accession number has RNA-seq quantifications
available. It does this by checking if the GEO accession number has a
"RNA-Seq raw counts" link available on the GEO download page.
}
\examples{
hasRNASeqQuantifications("GSE164073")

}
