% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseGEO.R
\name{parseGDS}
\alias{parseGDS}
\title{parse a GEO dataset (GDS)}
\usage{
parseGDS(fname)
}
\arguments{
\item{fname}{the filename of the SOFT format file. May be gzipped.}
}
\description{
GEO datasets, or GDS, used to be produced by NCBI GEO as "curated" versions
of submitted data. NCBI GEO no longer produces these objects, but there are
still thousands available that, in some cases, have much nicer sample
annotation and more standardized annotation for the associated GPLs.
}
\keyword{internal}
