% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseGEO.R
\name{parseGSEMatrix}
\alias{parseGSEMatrix}
\title{Parse a GSE mstrix file}
\usage{
parseGSEMatrix(fname, AnnotGPL = FALSE, destdir = tempdir(),
  getGPL = TRUE, parseCharacteristics = TRUE)
}
\arguments{
\item{fname}{filename}

\item{AnnotGPL}{set to TRUE to get the annotation GPL version}

\item{destdir}{the destdination directory for download}

\item{getGPL}{whether or not to get the GPL associated}

\item{parseCharacteristics}{Whether or not to do full "characteristic" parsing}
}
\description{
Not meant for user calling, but parses
a single GSEMatrix file.
}
\keyword{internal}
