% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{rnaseq_quantification_results}
\alias{rnaseq_quantification_results}
\title{Get RNA-seq quantification and annotation from GEO}
\usage{
rnaseq_quantification_results(gse)
}
\arguments{
\item{gse}{GEO accession number}
}
\value{
A list with two elements: quants (a matrix of raw counts) and
annotation (a data frame of annotation information).
}
\description{
This function downloads the raw counts and annotation files from GEO
for a given GEO accession number.
}
\examples{
res <- rnaseq_quantification_results("GSE83322")

# get dimensions of the raw counts and annotation
sapply(res, dim)

# check the structure of the results
sapply(res, class)

# check the structure of the results
head(res$quants)
head(res$annotation)

nrows(res$quants) == nrow(res$annotation)

}
\keyword{internal}
