% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchGEO.R
\name{searchFieldsGEO}
\alias{searchFieldsGEO}
\title{Provide a list of possible search fields for GEO search}
\usage{
searchFieldsGEO()
}
\value{
a data.frame with names of possible search fields for GEO search
as well as descriptions, data types, etc. for each field. Fields are
in rows and their properties are in columns.
}
\description{
Provide a list of possible search fields for GEO search
}
\examples{
searchFieldsGEO()

}
\seealso{
\code{\link{searchGEO}}
}
