% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{urlExtractRNASeqQuantGenomeInfo}
\alias{urlExtractRNASeqQuantGenomeInfo}
\title{Extract genome build and species from a GEO download URL}
\usage{
urlExtractRNASeqQuantGenomeInfo(url)
}
\arguments{
\item{url}{A GEO annotation file download URL}
}
\value{
A character vector with the genome build and species information
}
\description{
This function extracts the genome build and species information
from a GEO download URL. The genome build and species information
is expected to be in the filename of the download URL.
}
\keyword{internal}
