% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browseGEOAccession.R
\name{urlForAccession}
\alias{urlForAccession}
\title{The URL for a GEO accession}
\usage{
urlForAccession(geo)
}
\arguments{
\item{geo}{A GEO accession number}
}
\value{
A character vector with the URL for the GEO accession
}
\description{
Sometimes, you just need the URL for a GEO accession. This function
returns the URL for a given GEO accession number that
can be used to access the GEO page for that accession.
}
\examples{

urlForAccession("GSE262484")

}
\seealso{
\code{\link{browseGEOAccession}}
}
