% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEOquery-package.R
\name{GEOData-accessors}
\alias{GEOData-accessors}
\alias{dataTable}
\alias{Accession}
\alias{Columns}
\alias{GPLList}
\alias{GSMList}
\alias{Meta}
\alias{Table}
\title{Generic functions for GEOquery}
\description{
The main documentation is in the Class documentation
}
\seealso{
\code{\link{GEOData-class}}
}
\author{
Sean Davis
}
\keyword{IO}
