% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEOquery-package.R
\docType{class}
\name{GEODataTable-class}
\alias{GEODataTable-class}
\alias{Accession,GEODataTable-method}
\alias{Columns,GEODataTable-method}
\alias{Meta,GEODataTable-method}
\alias{Table,GEODataTable-method}
\alias{dataTable,GEODataTable-method}
\alias{show,GEODataTable-method}
\title{Class "GEODataTable"}
\description{
Contains the column descriptions and data for the datatable part of a GEO
object
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("GEODataTable", ...)}.
}

\author{
Sean Davis
}
\keyword{classes}
