% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGEOSuppFiles.R
\name{getGEOSeriesFileListing}
\alias{getGEOSeriesFileListing}
\title{GSE Supplemental file listing}
\usage{
getGEOSeriesFileListing(GSE)
}
\arguments{
\item{GSE}{character(1) the GSE accession}
}
\value{
A data.frame with 5 columns. See example.
}
\description{
The GEO Series records often have one or more supplemental files.
In most cases, those files are archived as '.tar' files, the contents
of which are only available in a file listing file not present on the
website for download.
}
\details{
This function reads that file listing file and returns the results
as a data.frame.
}
\examples{
\dontrun{
getGEOSeriesFileListing('GSE288770')

}
}
