% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGEOSuppFiles.R
\name{getGEOSuppFiles}
\alias{getGEOSuppFiles}
\title{Get Supplemental Files from GEO}
\usage{
getGEOSuppFiles(GEO, makeDirectory = TRUE, baseDir = getwd())
}
\arguments{
\item{GEO}{A GEO accession number such as GPL1073 or GSM1137}

\item{makeDirectory}{Should a "subdirectory" for the downloaded files be
created?  Default is TRUE.  If FALSE, the files will be downloaded directly
into the baseDir.}

\item{baseDir}{The base directory for the downloads.  Default is the current
working directory.}
}
\value{
A data frame is returned invisibly with rownames representing the
full path of the resulting downloaded files and the records in the
data.frame the output of file.info for each downloaded file.
}
\description{
NCBI GEO allows supplemental files to be attached to GEO Series (GSE), GEO
platforms (GPL), and GEO samples (GSM).  This function "knows" how to get
these files based on the GEO accession.  No parsing of the downloaded files
is attempted, since the file format is not generally knowable by the
computer.
}
\details{
Again, just a note that the files are simply downloaded.
}
\examples{

# a <- getGEOSuppFiles('GSM1137')
# a

}
\author{
Sean Davis <sdavis2@mail.nih.gov>
}
\keyword{IO}
\keyword{database}
