% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGEOSuppFiles.R
\name{getGEOSuppFileURL}
\alias{getGEOSuppFileURL}
\title{Get GEO supplemental file URL for a given GEO accession}
\usage{
getGEOSuppFileURL(GEO)
}
\arguments{
\item{GEO}{}
}
\description{
Get GEO supplemental file URL for a given GEO accession
}
\examples{
# an example of a GEO supplemental file URL
# with a set of single-cell RNA-seq data
url = getGEOSuppFileURL("GSE161228")
url

\dontrun{
  browseURL(url)
}

}
