% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{getRNASeqQuantResults}
\alias{getRNASeqQuantResults}
\title{Get RNA-seq quantification and annotation from GEO}
\usage{
getRNASeqQuantResults(gse)
}
\arguments{
\item{gse}{GEO accession number}
}
\value{
A list with two elements: quants (a matrix of raw counts) and
annotation (a data frame of annotation information).
}
\description{
This function downloads the raw counts and annotation files from GEO
for a given GEO accession number.
}
\keyword{internal}
