% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEOquery-package.R
\docType{class}
\name{GPL-class}
\alias{GPL-class}
\alias{GPL}
\alias{GPL,GDS-method}
\title{Class "GPL"}
\description{
Contains a full GEO Platform entity
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("GPL", ...)}.
}

\seealso{
\code{\link{GEOData-class}}
}
\author{
Sean Davis
}
\keyword{classes}
