\name{GPL-class}
\docType{class}
\alias{GPL-class}

\title{Class "GPL" }
\description{Contains a full GEO Platform entity}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("GPL", ...)}.
}
\section{Slots}{
	 \describe{
    \item{\code{dataTable}:}{Object of class \code{"GEODataTable"}}
    \item{\code{header}:}{Object of class \code{"list"} containing
      metadata associated with the GPL}
  }
}
\section{Extends}{
Class \code{"GEOData"}, directly.
}
\section{Methods}{
No methods defined with class "GPL" in the signature, but methods
applicable to GEOData are also applicable here.
}
\author{ Sean Davis}
\seealso{
	\code{\link{GEOData-class}}
}
\keyword{classes}
