% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEOquery-package.R
\docType{class}
\name{GSE-class}
\alias{GSE-class}
\alias{GPLList,GSE-method}
\alias{GSMList,GSE-method}
\alias{Meta,GSE-method}
\title{Class "GSE"}
\description{
Contains a GEO Series entity
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("GSE", ...)}.
}

\seealso{
\code{\link{GPL-class}},\code{\link{GSM-class}}
}
\author{
Sean Davis
}
\keyword{classes}
