% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEOquery-package.R
\docType{class}
\name{GSM-class}
\alias{GSM-class}
\title{Class "GSM"}
\description{
A class containing a GEO Sample entity
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("GSM", ...)}.
}

\seealso{
\code{\link{GEOData-class}}
}
\author{
Sean Davis
}
\keyword{classes}
