\name{GSM-class}
\docType{class}
\alias{GSM-class}

\title{Class "GSM"}
\description{A class containing a GEO Sample entity}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("GSM", ...)}.
}
\section{Slots}{
	 \describe{
    \item{\code{dataTable}:}{Object of class \code{"GEODataTable"}}
    \item{\code{header}:}{Object of class \code{"list"} containing the
      metadata associated with the sample}
  }
}
\section{Extends}{
Class \code{"GEOData"}, directly.
}
\section{Methods}{
No methods defined with class "GSM" in the signature, but methods that
apply to the \code{GEOData} also apply here.
}
\author{ Sean Davis}
\seealso{
	\code{\link{GEOData-class}}
}
\keyword{classes}
