% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGEOfile.R
\name{gunzip}
\alias{gunzip}
\title{Gunzip a file}
\usage{
gunzip(filename, destname = gsub("[.]gz$", "", filename), overwrite = FALSE,
  remove = TRUE, BFR.SIZE = 1e+07)
}
\arguments{
\item{filename}{The filename to be unzipped}

\item{destname}{The destination file}

\item{overwrite}{Boolean indicating whether or not to overwrite a destfile
of the same name}

\item{remove}{Boolean indicating whether or not to remove the original file
after completion}

\item{BFR.SIZE}{The size of the read buffer....}
}
\value{
Invisibly, the number of bytes read.
}
\description{
gunzip a file
}
\details{
This function was stripped out of R.utils due to breaking some stuff on the
bioconductor build machine.
}
\seealso{
\code{\link{gzfile}}
}
\author{
Original author: Henrik Bengtsson
}
\keyword{IO}
